refactor: store repo 이동 및 변경으로 인한 수정
