+++ src/main/java/com/spartaclub/orderplatform/domain/order/application/OrderService.java
@@ -17,7 +17,7 @@ import com.spartaclub.orderplatform.domain.order.presentation.dto.PlaceOrderResp
 import com.spartaclub.orderplatform.domain.product.domain.entity.Product;
 import com.spartaclub.orderplatform.domain.product.infrastructure.repository.ProductRepository;
 import com.spartaclub.orderplatform.domain.store.domain.model.Store;
-import com.spartaclub.orderplatform.domain.store.infrastructure.repository.StoreRepository;
+import com.spartaclub.orderplatform.domain.store.domain.repository.StoreRepository;
 import com.spartaclub.orderplatform.global.application.security.UserDetailsImpl;
 import com.spartaclub.orderplatform.user.domain.entity.User;
 import jakarta.persistence.EntityNotFoundException;
