+++ src/main/java/com/spartaclub/orderplatform/domain/store/infrastructure/scheduler/StoreReviewScheduler.java
@@ -1,14 +1,12 @@
-package com.spartaclub.orderplatform.domain.store.application.service;
+package com.spartaclub.orderplatform.domain.store.infrastructure.scheduler;
 
 import com.spartaclub.orderplatform.domain.review.repository.ReviewRepository;
 import com.spartaclub.orderplatform.domain.store.domain.model.Store;
-import com.spartaclub.orderplatform.domain.store.infrastructure.repository.StoreRepository;
+import com.spartaclub.orderplatform.domain.store.domain.repository.StoreRepository;
 import jakarta.transaction.Transactional;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -32,8 +30,7 @@ public class StoreReviewScheduler {
 
         List<UUID> storeIds = reviews.stream().map(r -> (UUID) r[0]).toList();
 
-        Map<UUID, Store> storeMap = storeRepository.findAllById(storeIds).stream()
-            .collect(Collectors.toMap(Store::getStoreId, Function.identity()));
+        Map<UUID, Store> storeMap = storeRepository.findAllById(storeIds);
 
         for (Object[] review : reviews) {
             UUID storeId = (UUID) review[0];
